Move utils to src/
